+++ src/test/java/com/iexec/sms/authorization/AuthorizationServiceTests.java
@@ -18,11 +18,13 @@ package com.iexec.sms.authorization;
 
 import com.iexec.commons.poco.chain.ChainDeal;
 import com.iexec.commons.poco.chain.ChainTask;
+import com.iexec.commons.poco.chain.ChainTaskStatus;
 import com.iexec.commons.poco.chain.WorkerpoolAuthorization;
 import com.iexec.commons.poco.security.Signature;
 import com.iexec.commons.poco.tee.TeeUtils;
 import com.iexec.commons.poco.utils.BytesUtils;
-import com.iexec.commons.poco.utils.TestUtils;
+import com.iexec.commons.poco.utils.HashUtils;
+import com.iexec.commons.poco.utils.SignatureUtils;
 import com.iexec.sms.chain.IexecHubService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -42,6 +44,13 @@ import static org.mockito.Mockito.when;
 @ExtendWith(MockitoExtension.class)
 class AuthorizationServiceTests {
 
+    private static final String ENCLAVE_ADDRESS = "0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee";
+    private static final String DEAL_ID = "0x2222222222222222222222222222222222222222222222222222222222222222";
+    private static final String WORKER_ADDRESS = "0x87ae2b87b5db23830572988fb1f51242fbc471ce";
+    private static final String CHAIN_TASK_ID = "0x1111111111111111111111111111111111111111111111111111111111111111";
+    private static final String POOL_PRIVATE = "0xe2a973b083fae8043543f15313955aecee9de809a318656c1cfb22d3a6d52de1";
+    private static final String POOL_WRONG_SIGNATURE = "0xf869daaca2407b7eabd27c3c4c5a3f3565172ca7211ac1d8bfacea2beb511a4029446a07cccc0884c2193b269dfb341461db8c680a8898bb53862d6e48340c2e1b";
+
     @Mock
     IexecHubService iexecHubService;
 
@@ -52,8 +61,8 @@ class AuthorizationServiceTests {
     @Test
     void shouldBeAuthorizedOnExecutionOfTeeTaskWithDetails() {
         final ChainDeal chainDeal = getChainDeal();
-        final ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        final WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = getTeeWorkerpoolAuth();
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.of(chainDeal));
 
@@ -81,8 +90,8 @@ class AuthorizationServiceTests {
                 .poolOwner("0xc911f9345717ba7c8ec862ce002af3e058df84e4")
                 .tag(BytesUtils.EMPTY_HEX_STRING_32)
                 .build();
-        final ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        final WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = getTeeWorkerpoolAuth();
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.of(chainDeal));
 
@@ -93,7 +102,7 @@ class AuthorizationServiceTests {
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenGetTaskFailedWithDetails() {
-        final WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
+        final WorkerpoolAuthorization auth = getTeeWorkerpoolAuth();
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.empty());
 
         final Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
@@ -103,8 +112,8 @@ class AuthorizationServiceTests {
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenTaskNotActiveWithDetails() {
-        final WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
-        final ChainTask chainTask = TestUtils.getChainTask(UNSET);
+        final WorkerpoolAuthorization auth = getTeeWorkerpoolAuth();
+        final ChainTask chainTask = getChainTask(UNSET);
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
 
         final Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
@@ -114,9 +123,8 @@ class AuthorizationServiceTests {
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenGetDealFailedWithDetails() {
-        final ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        final WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
-        auth.setSignature(new Signature(TestUtils.POOL_WRONG_SIGNATURE));
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = getTeeWorkerpoolAuth();
 
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.empty());
@@ -129,9 +137,14 @@ class AuthorizationServiceTests {
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenPoolSignatureIsNotValidWithDetails() {
         final ChainDeal chainDeal = getChainDeal();
-        final ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        final WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
-        auth.setSignature(new Signature(TestUtils.POOL_WRONG_SIGNATURE));
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final Signature signature = new Signature(POOL_WRONG_SIGNATURE);
+        final WorkerpoolAuthorization auth = WorkerpoolAuthorization.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .workerWallet(WORKER_ADDRESS)
+                .enclaveChallenge(ENCLAVE_ADDRESS)
+                .signature(signature)
+                .build();
 
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.of(chainDeal));
@@ -163,10 +176,25 @@ class AuthorizationServiceTests {
     // endregion
 
     // region utils
-    ChainDeal getChainDeal() {
+    private ChainDeal getChainDeal() {
         return ChainDeal.builder()
                 .poolOwner("0xc911f9345717ba7c8ec862ce002af3e058df84e4")
                 .tag(TeeUtils.TEE_SCONE_ONLY_TAG)
+                .build();
+    }
+
+    private ChainTask getChainTask(ChainTaskStatus status) {
+        return ChainTask.builder().dealid(DEAL_ID).status(status).build();
+    }
+
+    private WorkerpoolAuthorization getTeeWorkerpoolAuth() {
+        final String hash = HashUtils.concatenateAndHash(WORKER_ADDRESS, CHAIN_TASK_ID, ENCLAVE_ADDRESS);
+        final Signature signature = SignatureUtils.signMessageHashAndGetSignature(hash, POOL_PRIVATE);
+        return WorkerpoolAuthorization.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .workerWallet(WORKER_ADDRESS)
+                .enclaveChallenge(ENCLAVE_ADDRESS)
+                .signature(signature)
                 .build();
     }
     // endregion